+++ addOns/authhelper/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Support for recorded scripts in the Authentication Tester.
 
+### Changed
+- Updated to depend on Zest add-on 48.8.0.
 
 ## [0.26.0] - 2025-06-20
 ### Added